C2PA-280: Remove dependency on rust-font-tools (#37)
* Initial rough cut at SFNT support.

* C2PA-307: Fix SFNT checksum calculation (#23)

Fixes the calculation of the SFNT checksum.

* C2PA-308: Fix content hashing & checksum update (#29)

Fix up the chunk-mapping code to produce a list with the correct offsets and CAI/other tagging, and also fix a latent bug generating the SFNT checksum. Verify by building C2PATOOL against this library, and verifying that adding, removing & modifying manifests works as expected:

- The C2PA content hash should always verify.
- The head table's checksumAdjustment should always be correct.

(The branch name is misleading, as this code supports both box **and** non-box hashing...)

* C2PA-273: Use Table trait (#30)

* Re-factor imports to use core explicitly, add debug logging.

* Tweak logging.

* Log boxen.

* Fix bug in table-re-layout, add locations tracing.

* Fix Cai/Other sense, and be-tighten head-table hashing (only exclude checksumAdjustment)

* bookmark of ideas

* kleen idea

* Need to use fresh values for the new checksum, not stale ones.

* Regenerate test artifact with working code.

* close loop on checksum, and then checkpoint before the chainsaw.

* Discard & simplify.

* Naming, "ostensible" cleanup.

* Fix ostensibles in WOFF, note on head definition.

* Patch up tests.

* Group len/checksum/write into trait

* Switch woff to trait

* Slight cleanup using write trait

* Overactive rename

* PR suggestions + snft comment fixups (woff to come)

* WOFF comments updated (+ fixes)

* Typos

* Missed a couple of extra comments

---------

Co-authored-by: Christopher Kush (Monotype) <christopher.kush@monotype.com>

* C2PA-305: Use of a `enum FontError` in the `font_io` handler. (#31)

Following examples in other areas of the library, this moves the font-related errors to the asset_handlers/font_io.rs module as:

enum FontError.

At this point, keeping the changes small with just the move and keeping the errors the same. This should facilitate a quicker change for the other story which picks up testing in more earnest to be unblocked.

I also noticed the library also has other places with this type of error structure and provides a wrap_*_err fn, so I did the same to try to stay consistent with the other parts of the library.

* Additional change to remove unused dep.

* C2PA-305: More detailed font errors (#33)

An extension to the previous PR which now converts all internal font_io/sft_io/woff_io calls to use FontError as the error type. And then all of the crate level trait implementations, the error is converted to a crate::error::Error.

I also took the time to delete unused errors and split errors that were being used generically into more detailed errors.

For the SaveError I spun out a separate error enum to scope the errors for more details on the save errors.

* C2PA-312: Not using fonttools, so no need to ignore its problems. (#36)

We had added two ignore-rules in order to let (downstream dependencies of) fonttools build; but as we no longer depend on that crate, we no longer need the rules.

* C2PA-303: Harden SfntIO test matrix (#34)

Adds unit & integration tests to increase coverage, and hardens some routines in response.

The associated JIRA issue includes ZIP files of code-coverage reports for the epic branch, and for this topic branch. Refer to Confluence to generate your own coverage reports.

---------

Co-authored-by: Christopher Kush (Monotype) <christopher.kush@monotype.com>
Co-authored-by: Jeff Vance <jeffrey.vance@monotype.com>